added Optional class
This implementation of an Optional datastructure is heavily inspired by Promises as
Promises are basically Optionals but /delayed/ ones. This takes the useful parts of
Promises and makes them instantaneous while keeping all of the names (.then, .catch,
.finally) but also adding a few new ones such as .observe_then and .observe_catch.
Another difference is that the values are not encapsulated and hidden away but can
be accessed using .value. This change was made because without the /delay/ this doesn't
make that much sense anymore.
The reason for hiding the state of the promise and the value it encapsulates is to
make people use promises the right way (using .then instead of trying to access the
values another way). This reasoning doesn't apply here as forcing .then to be used
is because of the time element of promises.

Functions like "get_value", "get_value_or" aren't implemented because access to the
value is provided directly and having T as a parameter breaks a lot of things from
the type systems perspective (covariance vs. contravariance breaks).

Maybe some more useful functions can get added in the future.