added compile-time-settings
compile-time-settings are (if tree shaking / similar optimizations are applied at
compile time) a way to enable / disable features at build time.

this is implemented by having a `features.ts` file which contains a bunch of settings
(which are typed btw), most of them simple boolean values.

other code can import this file and check for specific settings and if the code is
optimized at build time this should exclude unused features entirely.

at the moment the following things can be set:
- `DISABLE_SIGNUPS`
- `DISABLE_ACCOUNT_MODIFICATION`
- `DISABLE_DEVICE_IDS`
- `DISABLE_TRACKING`
- `LOGGING`
- `DISABLE_WEBINTERFACE`
- `DISABLE_MFA`
- `DISABLE_REGISTER_TOKENS`
- `DISABLE_LOGOUT`
- `DISABLE_REFRESH_TOKENS`

> most of these things don't do anything yet, this will come at some point